Move IApiService to abstractions
