+++ src/ReportPortal.Client/ReportPortal.Client.csproj
@@ -26,7 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Include="..\*.targets" Pack="true" PackagePath="build\" Visible="false"/>
-    <None Include="..\..\Logo.png" Pack="true" PackagePath="\" Visible="false"/>
+    <None Include="..\*.targets" Pack="true" PackagePath="build\" Visible="false" />
+    <None Include="..\..\Logo.png" Pack="true" PackagePath="\" Visible="false" />
   </ItemGroup>
 </Project>
